+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 
 - Remove node-sass and sass-loader dev dependencies
 - Remove Spectre Sass framework dependency and replace it with plain CSS using CSS variables
+- Remove electron-webpack dependency
+- Upgrade to Electron 13, which uses Node 14.17.0, V8 9.1, and Chromium 91
 
 ## 3.6.0 - 4 May 2021